fix validator gaps and extract shared validation module
