+++ .github/scripts/validate-cip.py
@@ -2,12 +2,13 @@
 """
 Validation script for CIP README.md files.
 Validates YAML headers and required sections.
+
+Rules shared with the CPS validator live in validation_common.py.
 """
 
 import sys
 import re
 import json
-import yaml
 from pathlib import Path
 from typing import Dict, List, Tuple, Optional
 
@@ -16,6 +17,25 @@ try:
 except ImportError:
     print("Error: jsonschema library is required. Install it with: pip install jsonschema", file=sys.stderr)
     sys.exit(1)
+
+from validation_common import (
+    parse_frontmatter,
+    _strip_fenced_code_blocks,
+    extract_h2_headers,
+    extract_h1_headers,
+    validate_line_endings,
+    validate_no_h1_headings,
+    humanize_schema_error,
+    validate_number_field,
+    validate_leading_zero_lines,
+    validate_unquoted_question_marks,
+    validate_authors_field,
+    validate_created_field,
+    validate_license_field,
+    validate_label_url_format,
+    validate_label_entries,
+    validate_directory_name,
+)
 
 
 # Required fields for CIP headers (in required order)
@@ -67,135 +87,6 @@ PATH_TO_ACTIVE_SUBSECTIONS = {
 SCHEMA_PATH = Path(__file__).parent.parent / 'schemas' / 'cip-header.schema.json'
 with open(SCHEMA_PATH, 'r', encoding='utf-8') as f:
     CIP_HEADER_SCHEMA = json.load(f)
-
-
-def parse_frontmatter(content: str) -> Tuple[Optional[Dict], Optional[str], Optional[List[str]], Optional[str]]:
-    """Parse YAML frontmatter from markdown content.
-
-    Returns:
-        Tuple of (frontmatter_dict, remaining_content, raw_lines, parse_error).
-        On success parse_error is None. When the '---' delimiters are present but
-        the enclosed YAML cannot be parsed (e.g. an impossible date like month
-        13), frontmatter is None and parse_error carries the underlying reason so
-        the caller can report it instead of a misleading "missing frontmatter".
-    """
-    # Check for frontmatter delimiters - must start with ---
-    if not content.startswith('---'):
-        return None, content, None, None
-
-    # Find the closing delimiter (--- on its own line)
-    lines = content.split('\n')
-    if lines[0] != '---':
-        return None, content, None, None
-
-    # Find the closing ---
-    end_idx = None
-    for i in range(1, len(lines)):
-        if lines[i] == '---':
-            end_idx = i
-            break
-
-    if end_idx is None:
-        return None, content, None, None
-
-    # Extract frontmatter (lines between the two --- markers)
-    frontmatter_lines = lines[1:end_idx]
-
-    # Preprocess: quote standalone '?' values (YAML interprets '?' as explicit key indicator)
-    processed_lines = []
-    for line in frontmatter_lines:
-        # Match lines like "CIP: ?" or "Category: ?" and quote the ?
-        if re.match(r'^[A-Za-z][A-Za-z -]*:\s+\?+\s*$', line):
-            line = re.sub(r':\s+(\?+)\s*$', r': "\1"', line)
-        processed_lines.append(line)
-
-    frontmatter_text = '\n'.join(processed_lines)
-
-    # Extract remaining content (everything after the closing ---)
-    remaining_lines = lines[end_idx + 1:]
-    remaining_content = '\n'.join(remaining_lines)
-
-    try:
-        frontmatter = yaml.safe_load(frontmatter_text)
-        if frontmatter is None:
-            return None, content, None, None
-        return frontmatter, remaining_content, frontmatter_lines, None
-    except yaml.YAMLError as e:
-        return None, content, None, str(e)
-    except ValueError as e:
-        # Catches invalid date values that YAML tries to parse (e.g., month 13)
-        return None, content, None, str(e)
-
-
-def _strip_fenced_code_blocks(content: str) -> str:
-    """Blank out lines inside fenced code blocks, preserving line count.
-
-    Recognises CommonMark fenced code blocks: an opening fence is a line
-    indented up to 3 spaces whose first non-whitespace run is 3+ backticks or
-    3+ tildes. For backtick fences the info string (rest of the line) must
-    not contain backticks — this excludes inline code like '```X``` is a ...'
-    which is not a valid fence. The closing fence must use the same character
-    and be at least as long as the opening, followed only by whitespace.
-
-    Lines inside a fence — including the fence lines themselves — are
-    replaced with '' so that line-based regex extraction skips them. An
-    unclosed fence treats all trailing lines as fenced, which is the safe
-    default for the validator (no false positives).
-    """
-    backtick_open_re = re.compile(r'^ {0,3}(`{3,})([^`]*)$')
-    tilde_open_re = re.compile(r'^ {0,3}(~{3,})')
-    result = []
-    in_fence = False
-    fence_char = None
-    fence_len = 0
-    for line in content.split('\n'):
-        if in_fence:
-            close_re = re.compile(
-                r'^ {0,3}(' + re.escape(fence_char) + r'{' + str(fence_len) + r',})\s*$'
-            )
-            if close_re.match(line):
-                in_fence = False
-                fence_char = None
-                fence_len = 0
-            result.append('')
-        else:
-            m_b = backtick_open_re.match(line)
-            m_t = tilde_open_re.match(line) if not m_b else None
-            if m_b:
-                in_fence = True
-                fence_char = '`'
-                fence_len = len(m_b.group(1))
-                result.append('')
-            elif m_t:
-                in_fence = True
-                fence_char = '~'
-                fence_len = len(m_t.group(1))
-                result.append('')
-            else:
-                result.append(line)
-    return '\n'.join(result)
-
-
-def extract_h2_headers(content: str) -> List[str]:
-    """Extract all H2 headers (##) from markdown content, ignoring fenced code blocks."""
-    h2_pattern = r'^##\s+(.+)$'
-    headers = []
-    for line in _strip_fenced_code_blocks(content).split('\n'):
-        match = re.match(h2_pattern, line)
-        if match:
-            headers.append(match.group(1).strip())
-    return headers
-
-
-def extract_h1_headers(content: str) -> List[str]:
-    """Extract all H1 headers (#) from markdown content, ignoring fenced code blocks."""
-    h1_pattern = r'^#\s+(.+)$'
-    headers = []
-    for line in _strip_fenced_code_blocks(content).split('\n'):
-        match = re.match(h1_pattern, line)
-        if match:
-            headers.append(match.group(1).strip())
-    return headers
 
 
 def extract_section_body(content: str, section_name: str) -> str:
@@ -250,46 +141,6 @@ def extract_h3_headers_under_section(content: str, section_name: str) -> List[st
     return h3_headers
 
 
-def validate_line_endings(file_path: Path) -> List[str]:
-    """Validate that file uses UNIX line endings (LF, not CRLF).
-
-    Returns:
-        List of error messages (empty if valid)
-    """
-    errors = []
-
-    try:
-        with open(file_path, 'rb') as f:
-            content_bytes = f.read()
-
-        if b'\r\n' in content_bytes:
-            errors.append("File uses Windows line endings (CRLF). Use UNIX line endings (LF) instead.")
-
-        # Check for standalone \r without \n (old Mac line endings)
-        content_without_crlf = content_bytes.replace(b'\r\n', b'')
-        if b'\r' in content_without_crlf:
-            errors.append("File uses old Mac line endings (CR). Use UNIX line endings (LF) instead.")
-    except Exception as e:
-        errors.append(f"Error checking line endings: {e}")
-
-    return errors
-
-
-def validate_no_h1_headings(content: str) -> List[str]:
-    """Validate that no H1 headings are present in the document.
-
-    Returns:
-        List of error messages (empty if valid)
-    """
-    errors = []
-
-    h1_headers = extract_h1_headers(content)
-    if h1_headers:
-        errors.append(f"H1 headings are not allowed. Found: {', '.join(h1_headers)}")
-
-    return errors
-
-
 def _validate_field_order(frontmatter: Dict) -> List[str]:
     """Validate that header fields appear in the correct order.
 
@@ -316,43 +167,6 @@ def _validate_field_order(frontmatter: Dict) -> List[str]:
             f"Got: {', '.join(actual_known)}"
         )
 
-    return errors
-
-
-def _validate_cip_field(frontmatter: Dict) -> List[str]:
-    """Friendly validation of the CIP number field.
-
-    Leading-zero detection is performed separately against raw lines.
-    """
-    errors = []
-    if 'CIP' not in frontmatter:
-        return errors
-    value = frontmatter['CIP']
-    if isinstance(value, bool):
-        errors.append(
-            f"'CIP' must be a positive integer (e.g., 30) or '?' if unassigned. Got: {value!r}"
-        )
-        return errors
-    if isinstance(value, int):
-        if value < 1:
-            errors.append(
-                f"'CIP' must be a positive integer (1 or greater) or '?' if unassigned. Got: {value}"
-            )
-        return errors
-    if isinstance(value, str):
-        if re.fullmatch(r'\?+', value):
-            return errors
-        if re.fullmatch(r'[1-9]\d*', value):
-            return errors
-        if re.fullmatch(r'0\d*', value):
-            return errors  # Leading-zero check handles this
-        errors.append(
-            f"'CIP' must be a positive integer (e.g., 30) or '?' if unassigned. Got: {value!r}"
-        )
-        return errors
-    errors.append(
-        f"'CIP' must be a positive integer or '?' if unassigned. Got: {value!r}"
-    )
     return errors
 
 
@@ -391,22 +205,6 @@ def _validate_status_field(frontmatter: Dict) -> List[str]:
     return errors
 
 
-def _validate_authors_field(frontmatter: Dict) -> List[str]:
-    """Friendly validation of Authors entries."""
-    errors = []
-    entries = frontmatter.get('Authors')
-    if not isinstance(entries, list):
-        return errors
-    pattern = re.compile(r'^.+\s+<.+>$')
-    for i, entry in enumerate(entries):
-        if not isinstance(entry, str) or not pattern.match(entry):
-            errors.append(
-                f"'Authors' entry {i+1}: must be in the form 'Name <email>'. "
-                f"Got: {entry!r}. Example: 'John Doe <john.doe@email.domain>'"
-            )
-    return errors
-
-
 def _validate_implementors_field(frontmatter: Dict) -> List[str]:
     """Friendly validation of the Implementors field.
 
@@ -441,48 +239,6 @@ def _validate_implementors_field(frontmatter: Dict) -> List[str]:
         f"'Implementors' must be 'N/A', '[]', or a list of 'Name <email-or-URI>' strings. "
         f"Got: {value!r}"
     )
-    return errors
-
-
-def _validate_created_field(frontmatter: Dict) -> List[str]:
-    """Friendly validation of the Created field."""
-    errors = []
-    if 'Created' not in frontmatter:
-        return errors
-    value = frontmatter['Created']
-    if hasattr(value, 'isoformat'):
-        value = value.isoformat()
-    if not isinstance(value, str) or not re.fullmatch(r'\d{4}-\d{2}-\d{2}', value):
-        errors.append(
-            f"'Created' must be an ISO 8601 date in YYYY-MM-DD form. "
-            f"Got: {frontmatter['Created']!r}. Example: '2024-05-21'"
-        )
-    return errors
-
-
-def _validate_license_field(frontmatter: Dict) -> List[str]:
-    """Friendly validation of the License field, with a did-you-mean hint on near-misses."""
-    errors = []
-    if 'License' not in frontmatter:
-        return errors
-    value = frontmatter['License']
-    if not isinstance(value, str):
-        return errors
-    valid = ['CC-BY-4.0', 'Apache-2.0']
-    if value in valid:
-        return errors
-    normalized = re.sub(r'\s+', '-', value).lower()
-    valid_normalized = {v.lower(): v for v in valid}
-    suggestion = valid_normalized.get(normalized)
-    if suggestion:
-        errors.append(
-            f"'License' must be one of: {', '.join(valid)}. "
-            f"Got: {value!r} (did you mean {suggestion!r}?)"
-        )
-    else:
-        errors.append(
-            f"'License' must be one of: {', '.join(valid)}. Got: {value!r}"
-        )
     return errors
 
 
@@ -602,108 +358,6 @@ def _validate_solution_to_format(frontmatter: Dict) -> List[str]:
     return errors
 
 
-def _validate_discussions_format(entries: list) -> List[str]:
-    """Validate that string Discussions entries follow the 'Label: URL' form.
-
-    String entries must be of the form 'Label: URL' (with whitespace after the
-    colon). Dict entries ({Label: URL}) are validated by the JSON schema.
-
-    Returns:
-        List of error messages (empty if valid)
-    """
-    errors = []
-    if not isinstance(entries, list):
-        return errors
-
-    pattern = re.compile(r'^[^:\s][^:]*:\s+https?://\S+')
-    for i, entry in enumerate(entries):
-        if isinstance(entry, str) and not pattern.match(entry):
-            errors.append(
-                f"'Discussions' entry {i+1}: must be in the form 'Label: URL'. "
-                f"Got: {entry!r}. "
-                f"Example: 'Original PR: https://github.com/cardano-foundation/CIPs/pull/123'"
-            )
-    return errors
-
-
-def _validate_label_entries(entries: list, field_name: str, label_prefixes: List[str]) -> List[str]:
-    """Validate semantic rules for labeled 'Label: URL' entries.
-
-    When a label matches one of the given prefixes followed by -NNNN (e.g., CIP-0030):
-    - URL must be a GitHub CIPs repository link (PR or merged document)
-    - Pull request URLs require '?' suffix on the number (candidate)
-    - Merged URLs must NOT have '?' suffix
-
-    Other labels are allowed with any valid URL.
-
-    Args:
-        entries: List of label-URL entries (strings 'Label: URL' or dicts {Label: URL})
-        field_name: Name of the field for error messages
-        label_prefixes: List of label prefixes to validate (e.g., ['CIP'], ['CIP', 'CPS'])
-
-    Returns:
-        List of error messages (empty if valid)
-    """
-    errors = []
-
-    if not isinstance(entries, list):
-        return errors
-
-    prefix_group = '|'.join(label_prefixes)
-    label_pattern = re.compile(rf'^({prefix_group})-\d+(\?)?$')
-    pr_pattern = re.compile(r'https://github\.com/cardano-foundation/CIPs/pull/\d+')
-    merged_pattern = re.compile(
-        rf'https://github\.com/cardano-foundation/CIPs/(tree|blob)/[^/]+/({prefix_group})-\d+'
-    )
-    github_cips_pattern = re.compile(
-        rf'^https://github\.com/cardano-foundation/CIPs/(pull/\d+|tree/[^/]+/({prefix_group})-\d+|blob/[^/]+/({prefix_group})-\d+)'
-    )
-
-    for i, entry in enumerate(entries):
-        if isinstance(entry, dict) and len(entry) == 1:
-            label, url = next(iter(entry.items()))
-        elif isinstance(entry, str):
-            match = re.match(r'^([^:]+):\s+(.+)$', entry)
-            if not match:
-                continue
-            label, url = match.groups()
-        else:
-            continue
-
-        label = label.strip()
-        url = url.strip()
-
-        label_match = label_pattern.match(label)
-        if not label_match:
-            continue  # Non-matching labels are allowed with any URL
-
-        ref_id = label_match.group(0).rstrip('?')
-        has_question_mark = label_match.group(2) == '?'
-
-        if not github_cips_pattern.match(url):
-            errors.append(
-                f"'{field_name}' entry {i+1}: label '{label}' requires a GitHub CIPs repository URL "
-                f"(pull request or merged document). Got: {url}"
-            )
-            continue
-
-        is_pr = pr_pattern.search(url) is not None
-        is_merged = merged_pattern.search(url) is not None
-
-        if is_pr and not has_question_mark:
-            errors.append(
-                f"'{field_name}' entry {i+1}: Pull request URL requires '?' suffix on the reference "
-                f"(use '{ref_id}?' instead of '{ref_id}' to indicate candidate status)"
-            )
-        elif is_merged and has_question_mark:
-            errors.append(
-                f"'{field_name}' entry {i+1}: Merged document should not have '?' suffix "
-                f"(use '{ref_id}' instead of '{ref_id}?' since this document is merged)"
-            )
-
-    return errors
-
-
 def _entry_url(entry) -> Optional[str]:
     """Extract URL from a Discussions entry (string 'Label: URL', plain URL, or {Label: URL})."""
     if isinstance(entry, dict) and len(entry) == 1:
@@ -734,96 +388,6 @@ def _validate_discussions_has_pr(entries) -> List[str]:
         "'https://github.com/cardano-foundation/CIPs/pull/<N>'"
     )
     return errors
-
-
-def _schema_error_label(error) -> str:
-    """Human label for the location of a schema error (e.g. 'Category' or 'Authors entry 2')."""
-    path = list(error.absolute_path)
-    if not path:
-        return 'header'
-    parts = []
-    for p in path:
-        parts.append(f"entry {p + 1}" if isinstance(p, int) else str(p))
-    return ' '.join(parts)
-
-
-def _schema_type_name(t) -> str:
-    """Friendly name for a JSON Schema type (or list of types)."""
-    mapping = {
-        'object': 'a mapping', 'array': 'a list', 'string': 'text',
-        'integer': 'an integer', 'number': 'a number',
-        'boolean': 'true/false', 'null': 'null',
-    }
-    if isinstance(t, list):
-        return ' or '.join(_schema_type_name(x) for x in t)
-    return mapping.get(t, str(t))
-
-
-def humanize_schema_error(error) -> str:
-    """Turn a jsonschema ValidationError into a friendly, actionable message.
-
-    Inspects the failed keyword (``error.validator``) and the schema node that
-    failed (which carries the field ``description``). Falls back to the
-    library's default message for anything unrecognised, so no failure mode is
-    ever swallowed — it just may be less pretty.
-    """
-    label = _schema_error_label(error)
-    validator = error.validator
-    node = error.schema if isinstance(error.schema, dict) else {}
-    description = node.get('description')
-
-    if validator == 'enum':
-        allowed = ', '.join(repr(v) for v in error.validator_value)
-        return f"'{label}' must be one of: {allowed}. Got: {error.instance!r}"
-
-    if validator == 'type':
-        return f"'{label}' must be {_schema_type_name(error.validator_value)}. Got: {error.instance!r}"
-
-    if validator == 'minItems':
-        n = error.validator_value
-        unit = 'entry' if n == 1 else 'entries'
-        return f"'{label}' must contain at least {n} {unit}. Got: {error.instance!r}"
-
-    if validator in ('oneOf', 'anyOf'):
-        msg = f"'{label}' does not match any accepted form. Got: {error.instance!r}."
-        if description:
-            msg += f" Expected: {description}"
-        return msg
-
-    if validator == 'additionalProperties':
-        properties = node.get('properties', {})
-        unexpected = (
-            [k for k in error.instance if k not in properties]
-            if isinstance(error.instance, dict) else []
-        )
-        unexpected_str = ', '.join(repr(k) for k in unexpected) if unexpected else 'unknown field(s)'
-        return (
-            f"Unknown header field(s): {unexpected_str}. "
-            f"Allowed fields: {', '.join(properties.keys())}"
-        )
-
-    if validator == 'required':
-        missing = [
-            r for r in error.validator_value
-            if isinstance(error.instance, dict) and r not in error.instance
-        ]
-        missing_str = ', '.join(repr(m) for m in missing) if missing else 'a required field'
-        return f"Missing required header field(s): {missing_str}"
-
-    if validator in ('minLength', 'maxLength'):
-        limit = error.validator_value
-        length = len(error.instance) if isinstance(error.instance, str) else '?'
-        bound = 'at least' if validator == 'minLength' else 'at most'
-        return f"'{label}' must be {bound} {limit} characters long. Got {length}: {error.instance!r}"
-
-    if validator == 'pattern':
-        msg = f"'{label}' has an invalid format. Got: {error.instance!r}."
-        if description:
-            msg += f" Expected: {description}"
-        return msg
-
-    # Fallback: keep the library's message but anchor it to the field.
-    return f"'{label}': {error.message}"
 
 
 def validate_header(frontmatter: Dict) -> List[str]:
@@ -870,19 +434,22 @@ def validate_header(frontmatter: Dict) -> List[str]:
         errors.append(f"Schema error: {e.message}")
 
     # Friendly per-field value checks (schema enforces only type/structure for these)
-    errors.extend(_validate_cip_field(frontmatter))
+    errors.extend(validate_number_field(frontmatter, 'CIP'))
     errors.extend(_validate_title_field(frontmatter))
     errors.extend(_validate_status_field(frontmatter))
-    errors.extend(_validate_authors_field(frontmatter))
+    errors.extend(validate_authors_field(frontmatter))
     errors.extend(_validate_implementors_field(frontmatter))
-    errors.extend(_validate_created_field(frontmatter))
-    errors.extend(_validate_license_field(frontmatter))
+    errors.extend(validate_created_field(frontmatter))
+    errors.extend(validate_license_field(frontmatter))
     errors.extend(_validate_solution_to_format(frontmatter))
 
     # Validate CIP/CPS label semantic rules on Discussions
     if 'Discussions' in frontmatter:
-        errors.extend(_validate_discussions_format(frontmatter['Discussions']))
-        errors.extend(_validate_label_entries(frontmatter['Discussions'], 'Discussions', ['CIP', 'CPS']))
+        errors.extend(validate_label_url_format(
+            frontmatter['Discussions'], 'Discussions',
+            'Original PR: https://github.com/cardano-foundation/CIPs/pull/123'
+        ))
+        errors.extend(validate_label_entries(frontmatter['Discussions'], 'Discussions', ['CIP', 'CPS']))
         errors.extend(_validate_discussions_has_pr(frontmatter['Discussions']))
 
     return errors
@@ -1041,30 +608,6 @@ def validate_header_whitespace(raw_lines: List[str]) -> List[str]:
     return errors
 
 
-def validate_unquoted_question_marks(raw_lines: List[str]) -> List[str]:
-    """Validate that no header field has an unquoted '?' value.
-
-    A bare '?' is invalid YAML (it denotes an explicit key), which breaks
-    GitHub's frontmatter rendering: the header table is not displayed and
-    Discussions links are not clickable. Quoted values like ``CIP: "?"``
-    are fine and do not match here.
-
-    Returns:
-        List of error messages (empty if valid)
-    """
-    errors = []
-    for line in raw_lines:
-        match = re.match(r'^([A-Za-z][A-Za-z -]*):\s+\?+\s*$', line)
-        if match:
-            field = match.group(1)
-            errors.append(
-                f"'{field}' has an unquoted '?' value; a bare '?' is invalid YAML and breaks "
-                f"GitHub's frontmatter rendering (header table / clickable Discussions links). "
-                f'Use {field}: "?" until a number is assigned, or the assigned number.'
-            )
-    return errors
-
-
 def _ref_folder_exists(repo_root: Path, prefix: str, number: int) -> bool:
     """Check whether a CIP/CPS folder exists at the repo root."""
     return (repo_root / f"{prefix}-{number:04d}").is_dir()
@@ -1127,52 +670,6 @@ def validate_solution_to(frontmatter: Dict, file_path: Path) -> List[str]:
     return errors
 
 
-def validate_directory_name(frontmatter: Dict, file_path: Path) -> List[str]:
-    """Validate that a CIP with an assigned number lives in a correctly-named directory.
-
-    For CIP number N, the parent directory must be 'CIP-NNNN' (zero-padded to 4 digits;
-    no truncation for numbers >= 10000). Unassigned CIPs ('?', '??', etc.) skip this check.
-    """
-    errors = []
-
-    # Regardless of the CIP field, a numbered directory must be zero-padded to
-    # 4 digits (5+ digits without a leading zero for numbers >= 10000).
-    dir_match = re.fullmatch(r'CIP-(\d+)', file_path.parent.name)
-    if dir_match:
-        digits = dir_match.group(1)
-        if len(digits) != 4 and (len(digits) < 4 or digits.startswith('0')):
-            errors.append(
-                f"Directory name '{file_path.parent.name}' is not zero-padded to 4 digits. "
-                f"Expected: 'CIP-{int(digits):04d}'"
-            )
-            return errors
-
-    cip_value = frontmatter.get('CIP')
-    if cip_value is None:
-        return errors  # Missing field is reported by header validation
-
-    # Skip unassigned CIPs ('?', '??', etc.)
-    if isinstance(cip_value, str) and cip_value.startswith('?'):
-        return errors
-
-    # Parse to integer; non-numeric strings are caught by header validation
-    try:
-        cip_num = int(cip_value)
-    except (ValueError, TypeError):
-        return errors
-
-    expected_dir = f"CIP-{cip_num:04d}"
-    actual_dir = file_path.parent.name
-
-    if actual_dir != expected_dir:
-        errors.append(
-            f"Directory name '{actual_dir}' does not match the CIP number {cip_num}. "
-            f"Expected: '{expected_dir}'"
-        )
-
-    return errors
-
-
 def is_cip_file(file_path: Path) -> bool:
     """Check if file path indicates a CIP document."""
     path_str = str(file_path)
@@ -1211,24 +708,16 @@ def validate_file(file_path: Path) -> Tuple[bool, List[str]]:
             errors.append("Missing or invalid YAML frontmatter (must start with '---' and end with '---')")
         return False, errors
 
-    # Check for leading zeros in CIP field (YAML loses this information)
-    if raw_lines:
-        for line in raw_lines:
-            m = re.match(r'^CIP:\s+(0\d+)', line)
-            if m:
-                stripped = m.group(1).lstrip('0') or '0'
-                errors.append(
-                    f"'CIP' number must not have leading zeros. Got: {m.group(1)!r}. "
-                    f"Use 'CIP: {stripped}', not 'CIP: {m.group(1)}'."
-                )
-                break
+    # Check for unquoted leading zeros in the CIP field (YAML loses this
+    # information when it parses the value as a number)
+    errors.extend(validate_leading_zero_lines(raw_lines, frontmatter, 'CIP'))
 
     if raw_lines:
-        errors.extend(validate_unquoted_question_marks(raw_lines))
+        errors.extend(validate_unquoted_question_marks(raw_lines, 'CIP'))
         errors.extend(validate_header_whitespace(raw_lines))
 
     # Validate the directory name matches the assigned CIP number
-    dir_errors = validate_directory_name(frontmatter, file_path)
+    dir_errors = validate_directory_name(frontmatter, file_path, 'CIP')
     errors.extend(dir_errors)
 
     header_errors = validate_header(frontmatter)
